requirements.txt
kafka-python==2.0.2
flask==2.3.3